Refactor desktop modules and fix channel loading
