+++ desktop/package.json
@@ -8,8 +8,9 @@
     "dev": "vite",
     "build": "tsc && vite build",
     "typecheck": "tsc --noEmit",
+    "check:file-sizes": "node ./scripts/check-file-sizes.mjs",
     "lint": "biome lint .",
-    "check": "biome check .",
+    "check": "biome check . && pnpm check:file-sizes",
     "format": "biome format --write .",
     "preview": "vite preview",
     "tauri": "tauri",
