+++ desktop/src-tauri/src/commands/identity.rs
@@ -1,0 +1,81 @@
+use nostr::{EventBuilder, JsonUtil, Kind, Tag, ToBech32};
+use tauri::State;
+
+use crate::{app_state::AppState, models::IdentityInfo, relay::relay_ws_url};
+
+#[tauri::command]
+pub fn get_identity(state: State<'_, AppState>) -> Result<IdentityInfo, String> {
+    let keys = state.keys.lock().map_err(|error| error.to_string())?;
+    let pubkey = keys.public_key();
+    let pubkey_hex = pubkey.to_hex();
+    let bech32 = pubkey
+        .to_bech32()
+        .map_err(|error| format!("bech32 encode failed: {error}"))?;
+    let display_name = if bech32.len() > 16 {
+        format!("{}…{}", &bech32[..10], &bech32[bech32.len() - 4..])
+    } else {
+        bech32
+    };
+
+    Ok(IdentityInfo {
+        pubkey: pubkey_hex,
+        display_name,
+    })
+}
+
+#[tauri::command]
+pub fn get_relay_ws_url() -> String {
+    relay_ws_url()
+}
+
+#[tauri::command]
+pub fn sign_event(
+    kind: u16,
+    content: String,
+    tags: Vec<Vec<String>>,
+    state: State<'_, AppState>,
+) -> Result<String, String> {
+    let keys = state.keys.lock().map_err(|error| error.to_string())?;
+
+    let nostr_tags = tags
+        .into_iter()
+        .map(|tag| Tag::parse(tag).map_err(|error| format!("invalid tag: {error}")))
+        .collect::<Result<Vec<_>, _>>()?;
+
+    let event = EventBuilder::new(Kind::Custom(kind), content)
+        .tags(nostr_tags)
+        .sign_with_keys(&keys)
+        .map_err(|error| format!("sign failed: {error}"))?;
+
+    Ok(event.as_json())
+}
+
+#[tauri::command]
+pub fn create_auth_event(
+    challenge: String,
+    relay_url: String,
+    state: State<'_, AppState>,
+) -> Result<String, String> {
+    let keys = state.keys.lock().map_err(|error| error.to_string())?;
+
+    let mut tags = vec![
+        Tag::parse(vec!["relay", &relay_url])
+            .map_err(|error| format!("relay tag failed: {error}"))?,
+        Tag::parse(vec!["challenge", &challenge])
+            .map_err(|error| format!("challenge tag failed: {error}"))?,
+    ];
+
+    if let Some(token) = state.configured_api_token.as_deref() {
+        tags.push(
+            Tag::parse(vec!["auth_token", token])
+                .map_err(|error| format!("auth token tag failed: {error}"))?,
+        );
+    }
+
+    let event = EventBuilder::new(Kind::Custom(22242), "")
+        .tags(tags)
+        .sign_with_keys(&keys)
+        .map_err(|error| format!("sign failed: {error}"))?;
+
+    Ok(event.as_json())
+}
