+++ desktop/src-tauri/src/commands/mod.rs
@@ -1,0 +1,13 @@
+mod agents;
+mod channels;
+mod identity;
+mod messages;
+mod profile;
+mod tokens;
+
+pub use agents::*;
+pub use channels::*;
+pub use identity::*;
+pub use messages::*;
+pub use profile::*;
+pub use tokens::*;
