+++ desktop/src-tauri/src/managed_agents/mod.rs
@@ -1,0 +1,9 @@
+mod discovery;
+mod runtime;
+mod storage;
+mod types;
+
+pub use discovery::*;
+pub use runtime::*;
+pub use storage::*;
+pub use types::*;
